Stars *MORE WORK ON THIS NEEDS TO BE DONE*
This is for the following commit:

Added the stars dynamically for each option

Configured the tabs to change based on the user’s rating option currently selected.

Have the tabs switch to each link

What still needs to be done for the stars:

Update the stars when the user clicks on them.

Store the user’s rating of the stars + differentiate between which rating system the user used.

NOTE: To not mess up the grabscores function, since this commit doesn't mention the function of the stars, I marked the ids of the stars as "p0...", while the slider's id for each option is "q0..."